Fixed dependency resolution in pipeline
Previously dependencies were determined by sorting
an array with a custom __lt__ method, but this turned
out to be incorrect. Now a slightly different list
based algorithm is being used which calculates it
correctly.

Furthermore added the option to specify stages
with a plus, as in 3+ generates stages 3 and above,
filtering all invalid ones.